requirements.txt
streamlit==1.27.2
google-generativeai==0.3.1
python-dotenv==1.0.0
PyPDF2==3.0.1
docx2txt==0.8
python-pptx==0.6.21